djangorest/api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py36_0
beautifulsoup4=4.6.0=py36_0
bleach=1.5.0=py36_0
decorator=4.1.2=py36_0
django=1.11.3=py36_0
djangorestframework=3.6.4=py_0
entrypoints=0.2.3=py36_0
freetype=2.5.5=2
html5lib=0.999=py36_0
icu=54.1=0
ipykernel=4.6.1=py36_0
ipython=6.1.0=py36_0
ipython_genutils=0.2.0=py36_0
ipywidgets=6.0.0=py36_0
jedi=0.10.2=py36_2
jinja2=2.9.6=py36_0
jsonschema=2.6.0=py36_0
jupyter=1.0.0=py36_3
jupyter_client=5.1.0=py36_0
jupyter_console=5.2.0=py36_0
jupyter_core=4.3.0=py36_0
libiconv=1.14=0
libpng=1.6.30=1
libxml2=2.9.4=0
libxslt=1.1.29=0
lxml=3.8.0=py36_0
markupsafe=1.0=py36_0
mistune=0.7.4=py36_0
mkl=2017.0.3=0
nbconvert=5.2.1=py36_0
nbformat=4.4.0=py36_0
notebook=5.0.0=py36_0
numpy=1.13.1=py36_0
openssl=1.0.2l=0
pandas=0.20.3=py36_0
pandocfilters=1.4.2=py36_0
path.py=10.3.1=py36_0
pexpect=4.2.1=py36_0
pickleshare=0.7.4=py36_0
pip=9.0.1=py36_1
prompt_toolkit=1.0.15=py36_0
ptyprocess=0.5.2=py36_0
pygments=2.2.0=py36_0
pyqt=4.11.4=py36_2
python=3.6.2=0
python-dateutil=2.6.1=py36_0
pytz=2017.2=py36_0
pyzmq=16.0.2=py36_0
qt=4.8.7=4
qtconsole=4.3.1=py36_0
readline=6.2=2
requests=2.14.2=py36_0
selenium=3.5.0=py36_0
setuptools=27.2.0=py36_0
simplegeneric=0.8.1=py36_1
sip=4.18=py36_0
six=1.10.0=py36_0
sqlite=3.13.0=0
terminado=0.6=py36_0
testpath=0.3.1=py36_0
tk=8.5.18=0
tornado=4.5.2=py36_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
wheel=0.29.0=py36_0
widgetsnbextension=3.0.2=py36_0
xz=5.2.3=0
zlib=1.2.11=0
